+++ src/main/java/com/caffeine/gwanghwamun/common/jwt/JwtProperties.java
@@ -1,0 +1,16 @@
+package com.caffeine.gwanghwamun.common.jwt;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Getter
+@Setter
+@Configuration
+@ConfigurationProperties(prefix = "jwt")
+public class JwtProperties {
+
+	private String secretKey;
+	private long expirationTime;
+}
